feat(webui): open external sessions inline in native chat layout (#3217)
When a user clicks an external session from the sidebar "Ext" toggle,
the transcript now opens directly in the main chat content area — the
same unified layout used for native gptme conversations — rather than
navigating away to the separate /external-sessions page.

Changes:
- Extract ExternalSessionDetail as a standalone reusable component
  (transcript view + capability-gated steer_inject input) from the
  ExternalSessionsView two-panel catalog page
- ExternalSessionsView reuses the extracted component (no behaviour change)
- MainLayout reads the ?external=<id> URL param and renders
  ExternalSessionDetail in the main content area when set
- UnifiedSidebar navigates to /chat?external=<id> instead of
  /external-sessions?selected=<id> so the unified sidebar stays visible
- ConversationList highlights the selected external session in the sidebar
  (bg-accent ring-1 ring-primary, matching native conversation selection)

The /external-sessions route and two-panel catalog view are unchanged —
they remain accessible from the sidebar nav icon for the full catalog
experience.

All 567 frontend tests pass; TypeScript and ESLint clean.
