+++ webui/src/components/UnifiedSidebar.tsx
@@ -96,6 +96,9 @@ interface Props {
   // Split view
   onOpenInSplitView?: (conversationId: string) => void;
 
+  // External sessions
+  selectedExternalId?: string;
+
   // Task props
   tasks: Task[];
   selectedTaskId?: string;
@@ -126,6 +129,7 @@ export const UnifiedSidebar: FC<Props> = ({
   tasksError = false,
   onTasksRetry,
   onOpenInSplitView,
+  selectedExternalId,
 }) => {
   const selectedWorkspace = use$(selectedWorkspace$);
   const selectedAgent = use$(selectedAgent$);
@@ -151,9 +155,7 @@ export const UnifiedSidebar: FC<Props> = ({
 
   const handleSelectExternal = useCallback(
     (id: string) => {
-      const route = appRoute('/external-sessions');
-      const separator = route.includes('?') ? '&' : '?';
-      navigate(`${route}${separator}selected=${encodeURIComponent(id)}`);
+      navigate(`/chat?external=${encodeURIComponent(id)}`);
     },
     [navigate]
   );
@@ -427,6 +429,7 @@ export const UnifiedSidebar: FC<Props> = ({
               onOpenInSplitView={onOpenInSplitView}
               externalSessions={externalSessions}
               onSelectExternal={handleSelectExternal}
+              selectedExternalId={selectedExternalId}
             />
           )}
